Start an omnibox widget
This will be used for easy keyboard-based navigation and will also allow for
the entry of commands. While the whole UI won't be based around this, the
idea is that those who like a more CLI-like experience will be able to type
into the box, while those who prefer buttons and things can use those.