Add a base class for all navigation panes
I want the panes themselves to take care of passing focus to their content,
I don't want the navigation sidebar to have deep knowledge of what's inside
the panes.

So here I'm setting things up for an `activate` method that the navigation
sidebar can call on any of the panes and it can be confident they'll do the
right thing.