Add initial support for a configuration file
At the moment this only has a light/dark mode toggle. Note that I've added
that to the configuration rather than as some sort of runtime binding
because:

a. Folk seldom want to toggle light/dark; they have a preference.
b. If it's decided it does need a toggle, it should persist anyway.