Rework how the history pane is populated
I'd like the newest to be at the top, and the oldest at the bottom. I'm
using OptionList at the moment and right now the OptionList doesn't have any
sort of insert facility. It's starting to become clear it probably should.
For now though let's do the obvious thing and repopulate each time.