Capture RequestError in build_raw_forge_url
This pretty much turns it into a no-op. In this case that feels fair;
there's a problem with the connection and we're trying to make guesses on
behalf of the user; this is our guess: there's no repo to forge to speak to
right now and so no guess to be made.